requirements.txt
bitdotio[psycopg2-binary]
dnspython>=2.3.0
pandas>=1.5.3
